requirements.txt
asgiref==3.8.1
dj-database-url==3.0.0
Django==5.2.3
django-ckeditor==6.7.3
django-js-asset==3.1.2
gunicorn==23.0.0
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
whitenoise==6.9.0
